+++ torchtitan/components/quantization/utils.py
@@ -5,7 +5,6 @@
 # LICENSE file in the root directory of this source tree.
 
 import torch.nn as nn
-from torchtitan.config import JobConfig
 
 
 def module_filter_fn(mod: nn.Module, fqn: str, filter_fqns: list[str]) -> bool:
@@ -26,25 +25,3 @@ def module_filter_fn(mod: nn.Module, fqn: str, filter_fqns: list[str]) -> bool:
     is_filtered_fqn = any(filter_fqn in fqn for filter_fqn in filter_fqns)
 
     return dims_multiples_of_16 and not is_filtered_fqn
-
-
-def validate_quantization_converters(job_config: JobConfig):
-    """
-    Validates that the job config uses the same quantization type for dense and MoE layers.
-    """
-    # TODO: Explore supporting applying different quantization methods to dense and MoE layers.
-    # quantization converter format:
-    # `quantize.[dense|moe].[mx|float8]`
-    quantization_type = lambda converter: converter.split(".")[-1]
-    existing_quantization_converter = None
-    for converter in job_config.model.converters:
-        if "quantize" in converter:
-            if existing_quantization_converter is None:
-                existing_quantization_converter = converter
-            else:
-                assert quantization_type(converter) == quantization_type(
-                    existing_quantization_converter
-                ), (
-                    "Cannot combine model converters with different quantization types: "
-                    f"'{quantization_type(converter)}' and '{quantization_type(existing_quantization_converter)}'"
-                )
